src/requirements.txt
sds011lib
ipython
flask
redis
APScheduler==3.7.0
flask-cors
python-aqi==0.6.1

